+++ src/main/java/yapp/buddycon/hello/HelloController.java
@@ -3,7 +3,7 @@ package yapp.buddycon.hello;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import yapp.buddycon.web.member.auth.AuthMember;
+import yapp.buddycon.common.domain.AuthMember;
 
 @RestController
 @RequestMapping("/api/v1/hello")
