+++ src/main/java/yapp/buddycon/web/member/auth/adapter/in/response/TokenResponseDto.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.web.member.auth;
+package yapp.buddycon.web.member.auth.adapter.in.response;
 
 public record TokenResponseDto(
   String grantType,
